restructure middleware directory
